% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_var_to_positive.R
\name{convert_var_to_positive}
\alias{convert_var_to_positive}
\title{Converts a var model to a positive notation, returns the updated var model}
\usage{
convert_var_to_positive(var_model, negative_variables)
}
\arguments{
\item{var_model}{the var model that should be converted to positive}

\item{negative_variables}{the variables that are negative, and should be positivised}
}
\value{
the positivised var-model
}
\description{
Converts a var model to a positive notation, returns the updated var model
}

