% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{create_result_matrix}
\alias{create_result_matrix}
\title{Creates a square matrix of size names, with a default value and named rows/columns}
\usage{
create_result_matrix(names, default_value = 0)
}
\arguments{
\item{names}{the names to use for the column names}

\item{default_value}{the value to set in each field of the matrix (default 0)}
}
\value{
a length(names) by length(names) matrix
}
\description{
Creates a square matrix of size names, with a default value and named rows/columns
}

