% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{set_exo}
\alias{set_exo}
\title{This is terrible practice. Irf requires the exogen matrix to be available, which is not available anymore.
Here we recreate the matrix by padding the datamat with zeros.}
\usage{
set_exo(model)
}
\arguments{
\item{model}{the model for which to set the exogen matrix in the global space}
}
\description{
This is terrible practice. Irf requires the exogen matrix to be available, which is not available anymore.
Here we recreate the matrix by padding the datamat with zeros.
}

