% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_functions.R
\docType{class}
\name{VarsFunctions-class}
\alias{VarsFunctions}
\alias{VarsFunctions-class}
\title{Facade for the vars package.}
\description{
Facade for the vars package.
}
\section{Fields}{

\describe{
\item{\code{bootstrap_iterations}}{the number of bootstrap iterations to do for determining the significance of the effects}

\item{\code{horizon}}{the number of steps to look in the future}

\item{\code{var_model}}{the var model to perform the calculations on}

\item{\code{orthogonalize}}{use orthogonalized IRF}
}}

